backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
arxiv==2.2.0
astropy==7.1.1
astropy-iers-data==0.2025.10.20.0.39.8
astroquery==0.4.11
attrs==25.4.0
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
docstring_parser==0.17.0
elastic-transport==9.2.0
elasticsearch==9.1.1
fastapi==0.119.1
feedparser==6.0.12
fonttools==4.60.1
frozenlist==1.8.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-aiplatform==1.122.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.15.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.46.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.6.0
kiwisolver==1.4.9
matplotlib==3.10.7
more-itertools==10.8.0
multidict==6.7.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
-e git+https://github.com/rsalehin/project-kepler-elastic.git@8650d17e1d23a03eb3cfeb4a3aa9fcd342581076#egg=project_kepler_backend&subdirectory=backend
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pyerfa==2.0.1.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyvo==1.7.1
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
sgmllib3k==1.0.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
webencodings==0.5.1
websockets==15.0.1
yarl==1.22.0
